Release docs to a gh-pages branch
As part of the `build:and_release_if_updated` task also build a static
version of the templates with a couple of minor variants. This will let
us show people what the template looks like in its barest form.

Push these rendered templates to the gh-pages branch which will be
publicly viewable on github.com.

It also copies in a copy of all the pkg tar files so that people can
download a version of the compiled templates.