Force scrollbar to always display on IE10/11
The scrollbar used to be constantly visible pre-IE10. In IE10 and
later, the scrollbar only appears when the user tries scrolling
or hovers over the right side of the screen.
This is causing IE users confusion.
Major sites likes BBC, Amazon, Guardian have chosen to always display
the scrollbar in all versions of IE.
This commit ensures the scrollbar will be always visible for all
IE users on GOV.UK.