Remove inline JavaScript from the template
We'd like to enable Content Security Policy on GOV.UK, which provides
additional security against cross-site scripting and other vulnerabilities
in the site.

As part of this, we need to set an HTTP header to instruct the browser
to not execute any JavaScript that it encounters in <script> tags
on the page.

This is to defend against cases where somebody exploits GOV.UK and
we are not adequately protected against script tags injected in the
page by other means eg by a malicious author in a content
management system.

This commit moves inline JavaScript into asset files that can be
loaded from a whitelisted domain.

This commit has issues and should not be merged to master. If you're
reading this on master in the future, sorry. The ERB in ie-8.js is
not processed. Also, this will cause an extra 3 HTTP requests which
it would be good to mitigate if possible.