Make TemplateProcessor a generic base class
TemplateProcessor was implicitly a class which produced "plain" files - ie
files which are not templates. Renaming it makes it clearer what it generates
and it makes it consistent with the other processor classes which are named
after the format they generate, so let's make things consistent and treat all
output formats the same.