Test that every compiler supports every key in the templates
It is easy to miss adding support for every key that is yielded to in the erb
template to every compiler. In fact, unless you know you need to do this, you
could easily miss adding it to any.

The implementation is hacky, but should be robust against new compilers being
added.