Add a failing test for Play template parameter declaration bug
Play templates are almost like a function, so have a parameter declaration on
their first line. This means that we need to define all the parameters used in
the template at the top.

This test is a crude way of checking whether each parameter is mentioned in the
declaration. It doesn't check that it is syntactically valid or that it is
correct.

A better guard against the problem might be to have Continuous Integration
(Travis) test an example Play app against code changes, but that is a major
piece of work.