Add Rails integration test for compiled template
We want to be confident that the changes we make to govuk_template don't
break the compiled templates. The only way we can really do this automatically
is to have code that can process each template.

The spec is currently wired up to anything and is hardcoded for the filename
generated by the Rails integration code. The expectations should be portable,
however.

This needs some scripting to trigger the sub-build within each integration
testing app and then run the spec. For the languages that aren't Ruby, we'll
need to alter the travis.yml.