Update Crown Copyright link.
The original link was redirecting to the new address, so updating this.

Coincidentally they seem to have moved to `http` from `https`. The same content gets served at via `https`, but the javascript doesn’t load as it is served from an `http` address which causes their menu to appear.