Create new non-retina crown images
We’re going to insert these into the page with CSS, so IE8 and lower will need images that are the correct size. We could use a filter scale to resize them in browser, but next-to-no-one will be using IE8 on an HD display and IE filters can break links in strange ways.

I also ran the rest of the crowns through ImageOptim and managed to squeeze a few bytes off.