Add integration test for nunjucks renderer
This commit adds an integration test for rendering the *Jinja* GOV.UK Template
with the Nunjucks[1] template rendering engine. The test setup is based on the
Mustache inheritance one.

The two templating languages are mostly[2] compatible. Having a test will ensure
that we don’t get caught out by any of the incompatibilities.

This is to avoid having to maintain two separate, but identical versions of the
template for the two different languages, as discussed in [3].

1. https://mozilla.github.io/nunjucks/
2. http://mozilla.github.io/nunjucks/faq.html#can-i-use-the-same-templates-between-nunjucks-and-jinja2-what-are-the-differences
3. https://github.com/alphagov/govuk_template/pull/193