Add a hook for hidden content when JS is enabled
Some content we want to be hidden by default when JS is available,
eg a toggle, where the content is initially hidden.

Hiding it with JS will usually give a flash/re-layout, as there will
be a delay between the page initially loading and the JS executing
and hiding the content itself.

Instead, we can use the `js-enabled` class [1] that is set on `body`
before content loads, to conditionally hide `js-hidden` content.

We already do this is some other apps [2], [3], and we're going to
need it for the Toggle JS module we're going to use with Govspeak
attachments, so it makes sense to do it at a lower, more common,
level. govuk_template is also where `js-enabled` gets set, so it
makes sense to add the hook here too.

[1]: https://github.com/alphagov/govuk_template/blob/master/source/views/layouts/govuk_template.html.erb#L57
[2]: https://github.com/alphagov/govuk_elements/blob/a860e4d9e75f8bd390a80efb607778f95756ea53/public/sass/elements/_helpers.scss#L48-L51
[3]: https://github.com/alphagov/whitehall/blob/5631a1722e186b194f4f7bb1f53cd2eb56e48034/app/assets/stylesheets/frontend/global/_js-hidden.scss