Migrate viewport declaration from govuk_frontend_toolkit
The viewport declarations for Opera and IE used to be defined in govuk_frontend_toolkit. It would be better if that was purely mixins and functions, and it makes sense that these declarations should live in the same place as the meta viewport declaration.

I looked into removing them completely, but there’s conflicting evidence as to what versions of Opera uses them, and at least IE10’s snap mode on desktop needs an @viewport.