Fix skiplink for Voiceover on iOS (#213)
Voiceover on iOS behaves weirdly with our main skiplink. It highlights
it, which causes it to gain focus, and then it moves focus to a
seemingly random other target. I've witnessed it within a few reloads
select either a URL, or a paragraph item, or the Safari reload button,
or a previously accessed browser tab.

I've traced the root cause of this to be our implementation, which uses
`position: static`. This causes the element to become part of the main
page layout, which presumably confuses the screen reader's accessibility
tree.

Switching to a normal `left: 0` positioning implementation seems to fix
it. This causes the element to overlay the main logo in certain
situations, but any other position would potentially lead to this issue.
It's worth mentioning that other websites such as github.com or
webaim.org also take this approach.