Add NPM publisher for jinja package
New repo created at
  http://github.com/alphagov/govuk_template_jinja

and initially updated by running this code locally.

Follows the same pattern as other NPM published versions, like
mustache and ejs. There's quite a bit of duplication in the
publisher code, but this doens't feel like the time to refactor
it, partly as we're looking at what might replace govuk_template
as part of the Patters and Tools team discovery.