Create GemPublisher class - similar to other types
Treat the gem publishing action in the same way as for other
publishers.

This repo is natively a gem, and the other publishers are treated
as derivitives of that. Long term we want govuk_template to be
language/package agnostic and built to support many build outputs,
where `gem` is just one of them.

Treatin the gem publishing in the same way as other types is a small
step towards that, and will allow some refactoring of the `and_release_if_updated`
rake task, which has a lot of duplication