Remove generated gov.uk from relative print links
Relative links currently have a https://www.gov.uk prepended to them in the print view. This is fine for gov.uk but not for other consumers of the template. I’ve updated static to include the prefix (https://github.com/alphagov/static/pull/823); this commit removes it from the template.