Move long README sections to docs/ directory
The README has grown a lot over time, making it harder to read. And as a
single page adding additional doc only adds to the problem.

Split specific doc sections out into their own directory, for clarity. This
is a convention used on GOV.UK projects and encouraged in the README
styleguide: https://github.com/alphagov/styleguides/blob/master/use-of-READMEs.md

I've done this now because I want to add more detailed documentation on the
publishing process, and document how to fix missing published packages, which
is likely to be super long.

Hopefully this will encourage more detailed documentation on the repo in general.