Update README to link to commonly used packages
Moving packaging and publishing documentation to `docs` made it less
clear how to consume `govuk_template`. Add links to the common ways
to consume it.

It's still not super clear, but I don't think the original was either.

Ideally we should re-write the README/docs based on user needs of people
using the packages, based on research