Refactor README and split long documentation out (#235)
* Move long README sections to docs/ directory

The README has grown a lot over time, making it harder to read. And as a
single page adding additional doc only adds to the problem.

Split specific doc sections out into their own directory, for clarity. This
is a convention used on GOV.UK projects and encouraged in the README
styleguide: https://github.com/alphagov/styleguides/blob/master/use-of-READMEs.md

I've done this now because I want to add more detailed documentation on the
publishing process, and document how to fix missing published packages, which
is likely to be super long.

Hopefully this will encourage more detailed documentation on the repo in general.

* Cleanup CONTRIBUTING

- Provide more detail on how to run tests, so hopefully people do
- Update the inaccurate versioning advice and describe how we do
  version bumps now. It could still be improved a bit, but it's
  an improvement over what we had before.
- Fix some types/copy-paste-errors.

* Update README to link to commonly used packages

Moving packaging and publishing documentation to `docs` made it less
clear how to consume `govuk_template`. Add links to the common ways
to consume it.

It's still not super clear, but I don't think the original was either.

Ideally we should re-write the README/docs based on user needs of people
using the packages, based on research