Remove color within link attribute
In accessability testing it was flagged that a color attribute was
within a <link> attribute which is against the W3 HTML guidelines:
https://validator.w3.org/nu/?doc=https%3A%2F%2Fwww.gov.uk%2F

This was added in commit 0d300da1bdc470527a35a3501037672974741505 along
with other changes but without explanation about why the color part was
added specifically.

Removing this on an HTML page doesn't *seem* to make any significant
changes to the template.