Add docs for adding tabindex="-1" to fix the skiplink
The GOV.UK template sets a skiplink to `#content` (just tab on any
GOV.UK page and hit ENTER when "Skip to main content" is highlighted).

This allows users to navigate to the main content area using only the
keyboard. Pressing enter when focus is on the skiplink should then set
focus on the #content area.

In Safari, there's a bug where this doesn't happen unless
`tabindex="-1"` is set on the destination element (in this case `<main
id="content">`).
Here's the bug report:
https://bugs.chromium.org/p/chromium/issues/detail?id=37721

This pull request adds documentation to explain that `tabindex="-1"` is
required on the `<main>` element to fix this issue.