Replace use of gem publisher with local version
The Gem Publisher library runs git commands to tag
the version its publishing. It assumes creds for
this are already set up but, in the switch from
Jenkins to Travis, this is no longer the case.

This creates a local publisher, following the
interface of the ones we already have, and
wrapping our use of Gem Publisher.

This means we can set up the github creds in the
same way as the other publishers, using an
existing token.

The new publisher calls Gem Publisher so there's a
name conflict if it is also called GemPublisher.

Caveat: I'm no rubyist so there may be better
approaches to this but I figured following the
convention the other publishers had, of being
named after their templating language would at
least be sensible.