Adds rails stylesheet and javascript tag helpers
It adds the ability to understand and convert stylesheet_link_tag and
javascript_include_tag to html for the other languages in which our erb
template is being compiled to (e.g. Django, Mustache, Liquid, Jinja). We
do this because we want to use the SRI options that sprockets-rails 3.x
supports when using these templates with a rails app.