Logic to allow conditional loading of fonts.css
WHAT
This commit wraps logic around the processors to allow fonts to be
conditionally included using the new variable `fonts_stylesheets`

There is also a change to the `mustache_inheritance` processor as
it was incomplete and didnt generate a package.json which caused
npm to reject it when included in a project.

WHY
The `govuk_template` allows configuration for pretty much everything
except whether or not to include the fonts stylesheets. For a project
in GOV.UK Pay we need to be able to exclude fonts for style and
copyright reasons.