Show proposition menu toggle always
This fixes an accessibility issue whereby the screen reader on my iPhone
would become trapped on the hidden menu button when the menu was
expanded. It also seems sensible to allow the user to collapse the menu.

Previously the js-hidden class was used exclusively to hide the menu
button when the menu was expanded. The js-visible class seems to be used
to mean 'menu expanded', so I've re-used this class to implement the new
behaviour.