Add missing aria support to the proposition menu
The usage guidance for proposition links now includes aria labels to
better describe the 'Menu' button and proposition links.

The 'Menu' link is now a button to indicate it performs an action and
does not link to another site. The related Javascript now also toggles
the aria-expanded and aria-hidden attributes for the 'Menu' button and
links respectively.

Preiously when the site was viewed on narrow-width devices, the
'Serivice Name' links would appear to the left of the 'Menu' button but
come 2nd in the tab order, which was confusing. This is fixed by moving
the 'Service Name' link outside of the proposition links.