Switch menu button back to link
The original idea was to change the 'Menu' link to appear link the same
on the developer docs site, but this would involve removing the href
attribute which in turn breaks the JS code. Instead, I've reverted back
to a link but used the role attribute to avoid it being read as a link.