Make sure the default opengraph image is the fallback
Currently when an application sets a custom og:image (which it should),
the custom meta tag will be inserted below the default meta tag.

This means that Facebook still sees the big black crown as the primary
image and the user has to manually select a new image.

By flipping the meta tag and `:head` inserter, we make sure that the
custom image is the first one chosen for the page.