Merge font bundle into main bundle
Previously this bundle was separated since the fonts were Base64 encoded inline,
this meant having a separate file enabled caching on this file that changed infrequently.

Now we're using a more traditional approach, fonts are loaded on demand when used,
this means the declarations themselves are small enough to put in the main CSS bundle.

This reduces requests by one, meaning less roundtrips to the server which can be slow on mobile networks.