Don't apply styles to new cookie banner
We'll be testing a new cookie banner. These styles don't play nicely with the new banner, but we can't remove them yet, while the old banner is still being used (as the new banner is still in testing phase). Therefore, we only want to apply these styles to the old banner.

When the new banner has been made live, we can remove this code.