Add deprecation notice to GitHub pages
Adds a deprecation notice to GitHub pages, and also fix:

- the missing 'GOV.UK' text in the header
- the missing 'Crown copyright' in the footer
- the missing OGL license text
- missing top and bottom margins on the #main element

I have not fixed the cookie banner, because the GOV.UK Template documentation does not use cookies (and unfortunately there is no easy way to remove the banner entirely)

Note that index.html is normally generated by the DocsPublisher [1], based on data/index.yml, but that only happens when a new release gets published.

For that reason, the same changes have been manually applied to index.html. The update should get deployed to GitHub pages as soon as this is merged to the gh-pages branch, and the changes to data.yml should mean that if index.html is re-compiled as part of a future release they won't be overwritten.

[1]: https://github.com/alphagov/govuk_template/blob/3bc945d0abe82e6d43a542bc6d8c9088881e1636/build_tools/publisher/docs_publisher.rb#L32-L47